[DTV00869355] softap: zero clear the ACS channel loading informaiton when interface down
[Description]
Add a zero-clear clean-up for ACS channel loading information when
interface down.

The ACS information remains there even if the interface is down and
is cleared only before a new scan request. This causes an issue when
the supplicant attempts to start GO without scan, for example, with
commands below.

1. wpa_cli scan
2. Reload wpa_supplicant
   (driver remains loaded and ifconfig down/up only)
3. wpa_cli p2p_listen
4. wpa_cli p2p_peer <peer address>
5. wpa_cli p2p_connect <peer address> pbc

In step #5, our GO makes use of ACS informaiton obtained in step #1.
This patch clears it in step #2 and our GO starts itself on a channel
spcified by supplicant.

Feature: softap
Change-Id: I086199e82fa4ee65774ad28bc9f19fc50dfd7997
Signed-off-by: sho <sho.masuda@mediatek.com>
CR-Id: DTV00869355