[DTV00869017] misc: match BSSID and SSID if both are provided
[Description]
Add a SSID check, if provided, for BSSID based connection.

DLINK DIR-815 is a buggy dual band router that uses a same BSSID for
both 2.4G and 5G BSSes (different SSIDs).

When our driver tries to connect to the 5G BSS, it walks through
the internal BSS list and finds the 2.4G BSS first. Since our driver
currently doesn't check and match with SSID, it tries to connect to
2.4G BSS since it has the target BSSID.

This patch checks SSID (if supplied) and skips 2.4G so that we can
connect to 5G BSS.

Feature: misc
Change-Id: Ia6243a0d1497ff4ee6823ede48f1f6e81c44d902
Signed-off-by: sho <sho.masuda@mediatek.com>
CR-Id: DTV00869017